requirements.txt
autopep8==1.5.3
click==7.1.2
cycler==0.10.0
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.2.2
numpy==1.19.0
Pillow==7.1.2
pycodestyle==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
toml==0.10.1
https://download.pytorch.org/whl/cpu/torch-1.1.0-cp37-cp37m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.3.0-cp37-cp37m-linux_x86_64.whl
Werkzeug==1.0.1
